Add mappers/continuous_color_mapper
