+++ bokehjs/src/coffee/models/mappers/index.ts
@@ -1,4 +1,5 @@
 export {CategoricalColorMapper} from "./categorical_color_mapper"
+export {ContinuousColorMapper}  from "./continuous_color_mapper"
 export {ColorMapper}            from "./color_mapper"
 export {LinearColorMapper}      from "./linear_color_mapper"
 export {LogColorMapper}         from "./log_color_mapper"
